Address terminal drop PR review findings
The PR review found concrete issues in adjacent agent-routing support: CLI telemetry exposed host identifiers, agent hooks cleared notifications only at workspace scope, agent routing state survived normal panel close, and the socket path resolver lived in the CLI monolith. This narrows those behaviors while keeping the terminal drop hint and routing invariant unchanged.

Constraint: Do not edit CI YAML; keep socket resolver extraction inside the existing CLI target rather than introducing a new SwiftPM package boundary.

Rejected: Add a new SwiftPM target for the resolver | too broad for this PR and inconsistent with the repo's current CLI helper-file pattern.

Confidence: high

Scope-risk: moderate

Directive: Keep notification clears surface-scoped whenever an agent hook has a surface id.

Tested: git diff --check; plutil -lint GhosttyTabs.xcodeproj/project.pbxproj; swift_file_length_budget.py

Not-tested: Local XCTest/build execution per repository and user policy
